+++ cmd/neofs-lens/internal/storage/get.go
@@ -31,7 +31,7 @@ func getFunc(cmd *cobra.Command, _ []string) error {
 		return fmt.Errorf("invalid address argument: %w", err)
 	}
 
-	storage, err := openEngine()
+	storage, err := openEngine(true)
 	if err != nil {
 		return err
 	}